% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkForMcols}
\alias{.checkForMcols}
\title{.checkForMcols}
\usage{
.checkForMcols(alns, mcols, func.nm = NULL, err.func = warning)
}
\arguments{
\item{alns}{GAlignments object containing metadata columns to check}

\item{mcols}{character(n) Name(s) of metadata columns to check for}

\item{func.nm}{character(1) The name of the function that called .checkForMcols}

\item{err.func}{function The error function to invoke if test fails,
e.g. warning or stop. (Default: warning)}
}
\description{
Internal function for checking required metadata columns
are present.
}
\author{
Helen Lindsay
}
