% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarFromMSA.R
\docType{methods}
\name{cigarFromMSA}
\alias{cigarFromMSA}
\alias{cigarFromMSA,MsaDNAMultipleAlignment-method}
\alias{cigarFromMSA,DNAStringSet-method}
\title{CIGAR from Multiple Sequence Alignment}
\usage{
cigarFromMSA(aln, ...)

\S4method{cigarFromMSA}{MsaDNAMultipleAlignment}(aln, ..., ref = NULL)

\S4method{cigarFromMSA}{DNAStringSet}(aln, ..., ref = NULL)
}
\arguments{
\item{aln}{A multiple sequence alignment}

\item{...}{Extra arguments}

\item{ref}{character(1) Name of sequence in aln to use as a
reference (Default: NULL)}

\item{ref}{character(1) name of the alignment in alns to use as a reference
(Default: NULL)}
}
\description{
Construct a CIGAR string from a multiple sequence
alignment
}
\examples{
aln <- DNAStringSet(c("AA--CC","AATTCC","AATTC-"))
names(aln) <- c("A","B","C")
cigarFromMSA(aln)
cigarFromMSA(aln, ref = "A")
cigarFromMSA(aln, ref = "B")
cigarFromMSA(aln, ref = "C")
}
\author{
Helen Lindsay
}
