% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapped_reference.R
\docType{methods}
\name{gappedRef}
\alias{gappedRef}
\alias{gappedRef,DNAString,GAlignments-method}
\title{gappedRef}
\usage{
gappedRef(reference, alignments, ...)

\S4method{gappedRef}{DNAString,GAlignments}(reference, alignments, ...)
}
\arguments{
\item{reference}{Ungapped reference sequence}

\item{alignments}{Sequences aligned to the reference sequence}

\item{...}{Extra arguments}
}
\description{
Given a set of sequences aligned to a reference,
introduce gaps in the reference sequence corresponding to insertion
locations in the aligned sequences.
TO DO: HOW ARE AMBIGUOUS ALIGNMENTS TREATED?
}
\author{
Helen Lindsay
}
