% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insertion_seqs.R
\docType{methods}
\name{getInsSeqs}
\alias{getInsSeqs}
\alias{getInsSeqs,GAlignments-method}
\title{getInsSeqs}
\usage{
getInsSeqs(alignments, ...)

\S4method{getInsSeqs}{GAlignments}(alignments, ...)
}
\description{
Get insertion sequences
}
\author{
Helen Lindsay
}
