% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigarFromMSA.R
\name{recodeAln}
\alias{recodeAln}
\title{recodeAln}
\usage{
recodeAln(aln, ref = NULL, ref.gap = "+", gap = NULL, nuc = "N",
  nuc.chars = NULL)
}
\arguments{
\item{aln}{(DNAStringSet)}

\item{ref}{(DNAString) the reference sequence (Default: NULL)}

\item{ref.gap}{(character(1)) Replace locations where the reference
sequence has a gap ("-") with ref.gap in alns}

\item{gap}{(character(1))}

\item{nuc}{(character(1)) Replace non-gap bases with nuc}

\item{nuc.chars}{(character(n)) Vector defining characters to replace with "nuc"}
}
\description{
Internal AlignmentTools function for recoding an alignment.
}
\examples{
recodeAln(DNAStringSet(c("AA--CC", "AATTC-")), ref = DNAString("AATT--"))
}
\author{
Helen Lindsay
}
